refactor: otherNotification 패키지를 primaryNotification으로 변경
