+++ android/2023-emmsale/app/src/main/res/layout/fragment_primary_notification.xml
@@ -11,7 +11,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/white"
-        tools:context=".presentation.ui.notificationBox.otherNotification.PrimaryNotificationFragment">
+        tools:context=".presentation.ui.notificationBox.primaryNotification.PrimaryNotificationFragment">
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
